Move parseUnit for better encaptulation
