+++ src/features/suggestions/parser/DecomposedAmount.ts
@@ -1,27 +1,6 @@
 import _ from "lodash";
-import { toUnit } from "../convert";
+import { parseUnit } from "../convert";
 import parse, { Measurement } from "./amount";
-
-function parseUnitText(unitText?: string) {
-  const words = _.words(_.lowerCase(unitText));
-  if (_.size(words) === 0) return "";
-  const first = words[0];
-
-  //TODO fix this
-  if ((["fluid", "fl"].includes(first)) && _.size(words) > 1) {
-    return first + " " + words[1];
-  }
-  if (first === "of" && _.size(words) > 2) {
-    return (words[1] + "-" + words[2]);
-  }
-  return first;
-}
-
-function from(unitText?: string) {
-  return {
-    toUnit: _.partial(toUnit, parseUnitText(unitText)),
-  }
-}
 
 function compose(quantityText: string | undefined, unitText: string) {
   return quantityText + " " + unitText;
@@ -30,7 +9,7 @@ function compose(quantityText: string | undefined, unitText: string) {
 function createMeasurement(measurement?: Measurement) {
   return {
     quantity: measurement?.quantity || 0,
-    unit: from(measurement?.unitText).toUnit(),
+    unit: parseUnit(measurement?.unitText),
     unitText: measurement?.unitText,
     amountWithUnitText: _.partial(compose, measurement?.quantityText),
   }
